dynamic model selection
Somewhat working dynamic model selection. The page starts at route "/" with the default selected_model_inference being equal to "no model".

A user selects a radio button to select which model they want to begin running inference on and the page redirects to /start_inference with a re render of index.html but with selected_model_inference = "Model 1", "Model 2",  or "Model 3",